pyproject.toml
[tool.poetry]
name = "tick_data_analytics"
version = "0.1.0"
description = ""
authors = ["Guillaume Bournique <guillaume.bournique@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
dbt-core = "1.8.6"
dbt-snowflake = "1.8.3"
h5py = "^3.12.1"
matplotlib = "^3.9.2"
pandas = "^2.2.3"
pyarrow = "^17.0.0"
python = ">=3.11,<3.12"
python-dotenv = "^0.19.0"
seaborn = "^0.13.2"
scipy = "^1.14.1"
snowflake-snowpark-python = "^1.23.0"

[tool.poetry.group.test.dependencies]
autoflake = "^2.3.1"
black = "^24.10.0"
coverage = "^7.6.3"
flake8 = "^7.1.1"
isort = "^5.13.2"
mypy = "^1.12.0"
pandas-stubs = "^2.2.3.241009"
pylint = "^3.3.1"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
requests-mock = "^1.12.1"
ruff = "^0.7.0"
sqlfluff = "3.1.1"
sqlfluff-templater-dbt = "3.1.1"
types-requests = "^2.32.0.20241016"
yamllint = "1.35.1"
boto3 = "^1.35.45"
botocore = "^1.35.45"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 89
